refactor: move vault logic and utils to separate files
